requirements.txt
asgiref==3.8.1
attrs==23.2.0
coverage==7.5.0
Django==5.0.4
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
